geekbot_cli/test_main.py
## test_main.py
import unittest
from unittest.mock import patch
from main import main
from exceptions import APIKeyNotFoundError, StandupException

class TestMain(unittest.TestCase):

    @patch('cli.CLI.start')
    def test_main_workflow(self, mock_start):
        """
        Test the main workflow by ensuring the CLI's start method is called.
        """
        # Simulate the main function call
        main()

        # Assert that the start method was called once
        mock_start.assert_called_once()

    @patch('config_manager.ConfigManager.get_api_key')
    def test_main_api_key_retrieval(self, mock_get_api_key):
        """
        Test the retrieval of the API key during the main workflow.
        """
        # Set up the return value for the mocked get_api_key method
        mock_get_api_key.return_value = 'test_api_key'

        # Simulate the main function call
        main()

        # Assert that the get_api_key method was called once
        mock_get_api_key.assert_called_once()

    @patch('config_manager.ConfigManager.get_api_key', side_effect=APIKeyNotFoundError)
    def test_main_api_key_not_found(self, mock_get_api_key):
        """
        Test the behavior when the API key is not found during the main workflow.
        """
        with self.assertRaises(APIKeyNotFoundError):
            # Simulate the main function call, which should raise an APIKeyNotFoundError
            main()

    @patch('cli.CLI.start', side_effect=StandupException)
    def test_main_standup_exception(self, mock_start):
        """
        Test the behavior when a StandupException is raised during the main workflow.
        """
        with self.assertRaises(StandupException):
            # Simulate the main function call, which should raise a StandupException
            main()

if __name__ == '__main__':
    unittest.main()
